scripts/requirements.txt
Pillow>=9.5.0
PyMuPDF>=1.22.5
opencv-python>=4.8.0
numpy>=1.24.0
